move registration activities to subdir
sign in - sign up - password recovery moved from /activities to /activities/user_registration
