+++ HypeChatApp/app/src/main/res/layout/activity_forgot.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activities.ForgotActivity"
+        tools:context=".activities.user_registration.ForgotActivity"
         android:background="@drawable/signin_background"
 >
