+++ HypeChatApp/app/src/main/res/layout/activity_sign_up.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activities.SignUpActivity"
+        tools:context=".activities.user_registration.SignUpActivity"
         android:background="@drawable/signin_background"
 >
